Split control_windows table into 2
Split into a table for tracking limit establishment (aka Phase I,
trial limit window) and a table for the control period (aka Phase II,
online, in-control). Break the FK relationship between windows and
samples. Instead limit establishment windows FK to instruments and
control windows FK to limit establishment. Periods can be used to
map/join onto the samples table. This way the relationship is not
fixed at 1:1 and windows may overlap.

One downside is that there's no obvious way to enforce the constraint
that limit establishment and control windows should be immediately
adjacent but *not* overlapping.