Redesign whole_unit_conformance_inspections
Previously, this was keeping a count for conformant and non-conformant
units in a sample. This was very convenient for calculations but it's
not the 'real' state of the world. In particular it's more or less
identical to how I want to count conformities/non-conformities in units,
where multiple units make up a sample.

So this switches to tracking conformance/non-conformance at the
whole-of-unit level. In particular it's necessary to insert an inspection
for every item individually, rather than one record for the whole batch.