Introduce specs
Starting with Montgomery's flow width example. The approach chosen is
to use Ruby minitest specs. This was chosen over options like pgtap
or pgunit so that, in theory, it could be applied to other databases
in future. It's also easier to install Ruby and Bundler than to set up
pgtap/pgunit.

[#8]